Removing redundant casts

+++ src/net45/WampSharp/Core/Cra/WampCraHelpers.cs
@@ -152,7 +152,7 @@ namespace WampSharp.Core.Cra
                     hashLength++;
                 }
                 int keyLength = dklen/hashLength;
-                if ((long) dklen > (0xFFFFFFFFL*hashLength) || dklen < 0)
+                if (dklen > (0xFFFFFFFFL*hashLength) || dklen < 0)
                 {
                     throw new ArgumentOutOfRangeException(nameof(dklen));
                 }
